Fix terminal issues
In execing psql, we currently only pass PATH, which means a number of
environment variables that psql references are omitted. Most notably,
PAGER, EDITOR, and especially TERM can change behavior in weird ways.

Since the idea behind the plugin is to simulate invoking psql directly
on one of your attached (or pseudo-attached) databases, just pass the
entire env to the spawned process, with a default PGAPPNAME if one is
not provided.